+++ R/projection.R
@@ -10,32 +10,29 @@ wrong_proj_warning <-
          "  projecting to '", llcrs, "'")
 
 # Check and potentially adjust projection of objects to be rendered =======
-checkAdjustProjection <- function(x, method = "bilinear") {
-
-  x <- switch(class(x)[1],
-              "RasterLayer" = rasterCheckAdjustProjection(x, method),
-              "RasterStack" = rasterCheckAdjustProjection(x, method),
-              "RasterBrick" = rasterCheckAdjustProjection(x, method),
-              "SpatialPointsDataFrame" = spCheckAdjustProjection(x),
-              "SpatialPolygonsDataFrame" = spCheckAdjustProjection(x),
-              "SpatialLinesDataFrame" = spCheckAdjustProjection(x),
-              "SpatialPoints" = spCheckAdjustProjection(x),
-              "SpatialPolygons" = spCheckAdjustProjection(x),
-              "SpatialLines" = spCheckAdjustProjection(x),
-              "sf" = sfCheckAdjustProjection(x),
-              "XY" = sfCheckAdjustProjection(x),
-              "sfc_POINT" = sfCheckAdjustProjection(x),
-              "sfc_MULTIPOINT" = sfCheckAdjustProjection(x),
-              "sfc_LINESTRING" = sfCheckAdjustProjection(x),
-              "sfc_MULTILINESTRING" = sfCheckAdjustProjection(x),
-              "sfc_POLYGON" = sfCheckAdjustProjection(x),
-              "sfc_MULTIPOLYGON" = sfCheckAdjustProjection(x),
-              "sfc_GEOMETRY" = sfCheckAdjustProjection(x),
-              "sfc_GEOMETRYCOLLECTION" = sfCheckAdjustProjection(x))
-
-  return(x)
-
+checkAdjustProjection <- function(x, ...) {
+  UseMethod("checkAdjustProjection")
 }
+
+checkAdjustProjection.Raster <- function(x, ...) {
+  rasterCheckAdjustProjection(x, ...)
+}
+
+checkAdjustProjection.SpatialPointsDataFrame <-
+  checkAdjustProjection.SpatialPolygonsDataFrame <-
+  checkAdjustProjection.SpatialLinesDataFrame <-
+  checkAdjustProjection.SpatialPoints <-
+  checkAdjustProjection.SpatialPolygons <-
+  checkAdjustProjection.SpatialLines <- function(x, ...) {
+    spCheckAdjustProjection(x)
+  }
+
+checkAdjustProjection.sf <-
+  checkAdjustProjection.sfc <-
+  checkAdjustProjection.sfg <- function(x, ...) {
+    sfCheckAdjustProjection(x)
+  }
+
 #
 #   if (class(x)[1] %in% c("RasterLayer", "RasterStack", "RasterBrick")) {
 #     x <- rasterCheckAdjustProjection(x)
@@ -53,7 +50,7 @@ checkAdjustProjection <- function(x, method = "bilinear") {
 
 
 # Project Raster* objects for mapView =====================================
-rasterCheckAdjustProjection <- function(x, method) {
+rasterCheckAdjustProjection <- function(x, method = "bilinear") {
 
   is.fact <- raster::is.factor(x)[1]
 
@@ -94,10 +91,17 @@ starsCheckAdjustProjection <- function(x, method) {
   #     method = "ngb")
   #   x <- raster::as.factor(x)
   # } else {
-  x <- sf::st_transform(
-    x,
-    crs = llcrs
-  )
+
+  if (utils::packageVersion("stars") >= "0.4-1") {
+    x = stars::st_warp(x, crs = 3857)
+  } else {
+    x = sf::st_transform(x, crs = 3857)
+  }
+
+  # x <- sf::st_transform(
+  #   x,
+  #   crs = llcrs
+  # )
   # }
 
   return(x)