+++ R/sf.R
@@ -94,10 +94,16 @@ leaflet_sf <- function(x,
     if (inherits(sf::st_geometry(x), "sfc_MULTIPOINT"))
       x = suppressWarnings(sf::st_cast(x, "POINT"))
 
-    if (isTRUE(popup)) popup = leafpop::popupTable(x)
+    if (isTRUE(popup)) {
+      popup = leafpop::popupTable(x, className = "mapview-popup")
+    }
     if (inherits(popup, "character") &&
         all(popup %in% colnames(x))) {
-      popup = leafpop::popupTable(x, zcol = popup)
+      popup = leafpop::popupTable(
+        x
+        , zcol = popup
+        , className = "mapview-popup"
+      )
     }
 
     if (is.null(label)) label = makeLabels(x)
@@ -138,7 +144,7 @@ leaflet_sf <- function(x,
         x = suppressWarnings(sf::st_cast(x, "POINT"))
       }
       if (!(is.null(attributes(popup))) && names(attributes(popup)) == "popup") {
-        popup = leafpop::popupTable(x)
+        popup = leafpop::popupTable(x, className = "mapview-popup")
       }
     }
 
@@ -205,19 +211,19 @@ leafgl_sf = function(x,
   if (inherits(sf::st_geometry(x), "sfc_MULTIPOLYGON")) {
     x = suppressWarnings(sf::st_cast(x, "POLYGON"))
     if (!(is.null(attributes(popup))) && names(attributes(popup)) == "popup") {
-      popup = leafpop::popupTable(x)
+      popup = leafpop::popupTable(x, className = "mapview-popup")
     }
   }
   if (inherits(sf::st_geometry(x), "sfc_MULTILINESTRING")) {
     x = suppressWarnings(sf::st_cast(x, "LINESTRING"))
     if (!(is.null(attributes(popup))) && names(attributes(popup)) == "popup") {
-      popup = leafpop::popupTable(x)
+      popup = leafpop::popupTable(x, className = "mapview-popup")
     }
   }
   if (inherits(sf::st_geometry(x), "sfc_MULTIPOINT")) {
     x = suppressWarnings(sf::st_cast(x, "POINT"))
     if (!(is.null(attributes(popup))) && names(attributes(popup)) == "popup") {
-      popup = leafpop::popupTable(x)
+      popup = leafpop::popupTable(x, className = "mapview-popup")
     }
   }
 
@@ -361,12 +367,17 @@ leafgl_sf = function(x,
                    funs = funs,
                    args = args)
 
-  try(
-    if (attributes(popup)$popup == "leafpop") {
-      m$dependencies <- c(m$dependencies, popupLayoutDependencies())
-    }
-    , silent = TRUE
+  m$dependencies = c(
+    m$dependencies
+    , mapviewCSSDependencies()
   )
+
+  # try(
+  #   if (attributes(popup)$popup == "leafpop") {
+  #     m$dependencies <- c(m$dependencies, popupLayoutDependencies())
+  #   }
+  #   , silent = TRUE
+  # )
 
   if (is.function(legend)) m <- legend(m)
   m = removeDuplicatedMapDependencies(m)
@@ -524,6 +535,7 @@ leaflet_sfc <- function(x,
                         maxpoints,
                         attributes = NULL,
                         ...) {
+  ## remove geometry names (sfc-level)
   if (!is.null(names(x))) names(x) = NULL
   if (is_literally_false(highlight)) highlight = NULL
   if (is_literally_false(popup)) popup = NULL
@@ -680,12 +692,17 @@ leaflet_sfc <- function(x,
                    funs = funs,
                    args = args)
 
-  try(
-    if (attributes(popup)$popup == "leafpop") {
-      m$dependencies <- c(m$dependencies, popupLayoutDependencies())
-    }
-    , silent = TRUE
+  m$dependencies = c(
+    m$dependencies
+    , mapviewCSSDependencies()
   )
+
+  # try(
+  #   if (attributes(popup)$popup == "leafpop") {
+  #     m$dependencies <- c(m$dependencies, popupLayoutDependencies())
+  #   }
+  #   , silent = TRUE
+  # )
 
   # m$dependencies = c(
   #   m$dependencies